Replace old GroundMaterial with the new class
