+++ src/main/java/rs117/hd/data/materials/GroundMaterial.java
@@ -1,175 +1,84 @@
-/*
- * Copyright (c) 2021, 117 <https://twitter.com/117scape>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
 package rs117.hd.data.materials;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonSerializationContext;
-import com.google.gson.JsonSerializer;
-import java.lang.reflect.Type;
-import java.util.Arrays;
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonToken;
+import com.google.gson.stream.JsonWriter;
+import java.io.IOException;
 import java.util.Random;
 import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
+import rs117.hd.scene.GroundMaterialManager;
+import rs117.hd.utils.GsonUtils;
 
 @Getter
-@Deprecated
-public enum GroundMaterial {
-	NONE(Material.NONE),
-	SKULL_OBELISK(Material.SKULLS),
-	TRANSPARENT(Material.TRANSPARENT),
-	GRASS_1(Material.GRASS_1, Material.GRASS_2, Material.GRASS_3),
-	OVERWORLD_GRASS_1(Material.GRASS_1, Material.GRASS_2, Material.GRASS_3),
-	GRASS_SCROLLING(Material.GRASS_SCROLLING),
-	STONE_SCROLLING(Material.STONE_SCROLLING),
-	DIRT(Material.DIRT_1, Material.DIRT_2),
-	VERTICAL_DIRT(Material.DIRT_1_VERT, Material.DIRT_2_VERT),
-	SNOW_1(Material.SNOW_1, Material.SNOW_1, Material.SNOW_2, Material.SNOW_3, Material.SNOW_3, Material.SNOW_4),
-	SNOW_2(Material.SNOW_2, Material.SNOW_4),
-	GRAVEL(Material.GRAVEL),
-	FALADOR_PATHS(Material.FALADOR_PATH_BRICK),
-	VARROCK_PATHS(Material.JAGGED_STONE_TILE),
-	VARIED_DIRT(Material.GRAVEL, Material.DIRT_1, Material.DIRT_2),
-	VARIED_DIRT_SHINY(Material.GRAVEL_SHINY, Material.DIRT_1_SHINY, Material.DIRT_2_SHINY),
-	TILE_SMALL(Material.TILE_SMALL_1),
-	CARPET(Material.CARPET),
-	BRICK(Material.BRICK),
-	BRICK_BROWN(Material.BRICK_BROWN),
-	GRUNGE(Material.GRUNGE_1, Material.GRUNGE_2, Material.GRUNGE_1),
-	GRUNGE_2(Material.GRUNGE_2),
-	SUBMERGED_GRUNGE_2(Material.SUBMERGED_GRUNGE_2),
+@Slf4j
+public class GroundMaterial {
+	public static final GroundMaterial NONE = new GroundMaterial("NONE", Material.NONE);
+	public static final GroundMaterial DIRT = new GroundMaterial("DIRT", Material.DIRT_1, Material.DIRT_2);
+	public static final GroundMaterial UNDERWATER_GENERIC = new GroundMaterial(
+		"UNDERWATER_GENERIC",
+		Material.DIRT_1,
+		Material.DIRT_2
+	);
 
-	TILES_2x2_1(Material.TILES_2X2_1),
-	TILES_2x2_2(Material.TILES_2X2_2),
-	TILES_2X2_1_GLOSS(Material.TILES_2X2_1_GLOSS),
-	TILES_2X2_2_GLOSS(Material.TILES_2X2_2_GLOSS),
-	TILES_2X2_1_SEMIGLOSS(Material.TILES_2X2_1_SEMIGLOSS),
-	TILES_2X2_2_SEMIGLOSS(Material.TILES_2X2_2_SEMIGLOSS),
+	public String name;
+	public Material[] materials;
 
-	MARBLE_1(Material.MARBLE_1, Material.MARBLE_2, Material.MARBLE_3),
-	MARBLE_2(Material.MARBLE_3, Material.MARBLE_1, Material.MARBLE_2),
-	MARBLE_1_GLOSS(Material.MARBLE_1_GLOSS, Material.MARBLE_2_GLOSS, Material.MARBLE_3_GLOSS),
-	MARBLE_2_GLOSS(Material.MARBLE_3_GLOSS, Material.MARBLE_1_GLOSS, Material.MARBLE_2_GLOSS),
-	MARBLE_1_SEMIGLOSS(Material.MARBLE_1_SEMIGLOSS, Material.MARBLE_2_SEMIGLOSS, Material.MARBLE_3_SEMIGLOSS),
-	MARBLE_2_SEMIGLOSS(Material.MARBLE_3_SEMIGLOSS, Material.MARBLE_1_SEMIGLOSS, Material.MARBLE_2_SEMIGLOSS),
-	MARBLE_DARK(Material.MARBLE_DARK),
-
-	LASSAR_UNDERCITY_TILES(Material.LASSAR_UNDERCITY_TILES),
-	LASSAR_UNDERCITY_TILES_SUBMERGED(Material.LASSAR_UNDERCITY_TILES_SUBMERGED),
-	CAM_TORUM_TILES(Material.CAM_TORUM_TILES),
-
-	BLANK_SEMIGLOSS(Material.BLANK_SEMIGLOSS),
-
-	SAND(Material.SAND_1, Material.SAND_2, Material.SAND_3),
-
-	UNDERWATER_GENERIC(Material.DIRT_1, Material.DIRT_2),
-
-	WOOD_PLANKS_1(Material.WOOD_PLANKS_1),
-	CLEAN_WOOD_FLOOR(Material.CLEAN_WOOD_FLOOR),
-
-	HD_LAVA(
-		Material.HD_LAVA_1,
-		Material.HD_LAVA_2,
-		Material.HD_LAVA_1,
-		Material.HD_LAVA_1,
-		Material.HD_LAVA_2,
-		Material.HD_MAGMA_1,
-		Material.HD_MAGMA_2
-	),
-
-	STONE_PATTERN(Material.STONE_PATTERN),
-	CONCRETE(Material.CONCRETE),
-	SAND_BRICK(Material.SAND_BRICK),
-	CLEAN_TILE(Material.CLEAN_TILE),
-	WORN_TILES(Material.WORN_TILES),
-	WATER_FLAT(Material.WATER_FLAT),
-	HD_WOOD_PLANKS_2(Material.HD_WOOD_PLANKS_2),
-	ICE_1(Material.ICE_1),
-	ICE_1_HIGHGLOSS(Material.ICE_1_HIGHGLOSS),
-	ICE_4(Material.ICE_4),
-	WINTER_JAGGED_STONE_TILE(Material.WINTER_JAGGED_STONE_TILE),
-	WINTER_JAGGED_STONE_TILE_LIGHT(Material.WINTER_JAGGED_STONE_TILE_LIGHT),
-	WINTER_JAGGED_STONE_TILE_LIGHT_2(Material.WINTER_JAGGED_STONE_TILE_LIGHTER),
-	ROCKY_CAVE_FLOOR(Material.GRUNGE_2, Material.ROCK_2, Material.ROCK_2, Material.ROCK_1, Material.GRAVEL),
-	EARTHEN_CAVE_FLOOR(Material.GRUNGE_1, Material.DIRT_2, Material.DIRT_2, Material.ROCK_1, Material.DIRT_2),
-	STONE_CAVE_FLOOR(Material.STONE, Material.ROCK_1, Material.ROCK_2),
-	SANDY_STONE_FLOOR(Material.SAND_2, Material.STONE_NORMALED, Material.ROCK_2, Material.STONE_NORMALED),
-	PACKED_EARTH(Material.DIRT_1, Material.GRAVEL, Material.DIRT_1, Material.DIRT_1, Material.DIRT_2),
-	GRASSY_DIRT(Material.GRASS_1, Material.DIRT_1, Material.GRASS_2, Material.DIRT_2, Material.GRASS_3),
-	OOZE_FLOOR(Material.GRUNGE_1_SHINY, Material.GRAVEL_SHINY_LIGHT, Material.GRUNGE_2_SHINY),
-	TEMPLE_OF_THE_EYE_FLOOR(
-		Material.GRUNGE_1,
-		Material.GRAVEL_LIGHT,
-		Material.GRUNGE_2,
-		Material.GRAVEL_LIGHT
-	),
-	OVERWORLD_ROCKY(Material.ROCK_2, Material.STONE_NORMALED_DARK, Material.ROCK_1, Material.GRAVEL),
-	HD_BRICK_BROWN(Material.HD_BRICK_BROWN),
-	ABYSSAL_FLOOR(Material.ABYSSAL_2, Material.ABYSSAL, Material.ABYSSAL),
-	;
-
-	private final Material[] materials;
-
-	GroundMaterial(Material... materials) {
+	public GroundMaterial(String name, Material... materials) {
+		this.name = name;
 		this.materials = materials;
 	}
 
+	/**
+	 * Get a random material based on given coordinates.
+	 *
+	 * @param plane  The plane number.
+	 * @param worldX The X coordinate in world space.
+	 * @param worldY The Y coordinate in world space.
+	 * @return A randomly selected Material.
+	 */
 	public Material getRandomMaterial(int plane, int worldX, int worldY) {
-		// Generate a seed from the tile coordinates for
-		// consistent 'random' results between scene loads.
-		// This seed creates a patchy, varied terrain
+		// Generate a seed from the tile coordinates for consistent 'random' results between scene loads.
 		long seed = (long) (plane + 1) * 10 * (worldX % 100) * 20 * (worldY % 100) * 30;
 		Random randomTex = new Random(seed);
-		int randomInt = randomTex.nextInt(this.materials.length);
-		return this.materials[randomInt];
+		int randomIndex = randomTex.nextInt(materials.length);
+		return materials[randomIndex];
 	}
 
-	public static class GroundMaterialSerializer implements JsonSerializer<GroundMaterial> {
+	@Override
+	public String toString() {
+		return name;
+	}
+
+	@Slf4j
+	public static class JsonAdapter extends TypeAdapter<GroundMaterial> {
 		@Override
-		public JsonObject serialize(GroundMaterial groundMaterial, Type type, JsonSerializationContext jsonSerializationContext) {
-			JsonObject jsonObject = new JsonObject();
-			jsonObject.addProperty("name", groundMaterial.name());
+		public GroundMaterial read(JsonReader in) throws IOException {
+			if (in.peek() == JsonToken.NULL)
+				return null;
 
-			// Create materials array
-			JsonArray materialsArray = new JsonArray();
-			for (Material material : groundMaterial.getMaterials()) {
-				materialsArray.add(material.toString()); // Assuming Material has a toString() method
+			if (in.peek() == JsonToken.STRING) {
+				String name = in.nextString();
+				for (var groundMaterial : GroundMaterialManager.GROUND_MATERIALS)
+					if (name.equals(groundMaterial.name))
+						return groundMaterial;
+
+				log.warn("No ground material exists with the name '{}' at {}", name, GsonUtils.location(in), new Throwable());
+			} else {
+				log.warn("Unexpected type {} at {}", in.peek(), GsonUtils.location(in), new Throwable());
 			}
-			jsonObject.add("materials", materialsArray);
 
-			return jsonObject;
+			return null;
 		}
-	}
 
-	// Method to convert enum to JSON format
-	public static String toJson() {
-		Gson gson = new GsonBuilder()
-			.registerTypeAdapter(GroundMaterial.class, new GroundMaterialSerializer())
-			.setPrettyPrinting()
-			.create();
-		return gson.toJson(Arrays.asList(GroundMaterial.values()));
+		@Override
+		public void write(JsonWriter out, GroundMaterial groundMaterial) throws IOException {
+			if (groundMaterial == null) {
+				out.nullValue();
+			} else {
+				out.value(groundMaterial.name);
+			}
+		}
 	}
 }
